requirements.txt
mysql-connector-python~=8.0.32
pandas~=1.5.3
matplotlib~=3.7.1
numpy~=1.24.2